+++ src/symengine_assert.h
