+++ src/symengine_config.h.in
